+++ book-store/server/API/books/booksCont.ts
@@ -1,5 +1,5 @@
 import express from 'express'
-import connection from '../DB/database'
+import connection from '../../DB/database'
 
 export async function getAllBooks(req: express.Request, res: express.Response) {
     try {
