+++ book-store/server/server.ts
@@ -9,7 +9,7 @@ const PORT = process.evn.PORT || 4000
 //all app.use() are middlewares
 app.use(express.json())  
 
-import bookRoutes from "./API/booksRoutes"
+import bookRoutes from "./API/books/booksRoutes"
 app.use("/api/books", booksRoutes)
 
 app.listen(PORT, () => {
